# server/docs/swagger.yaml
definitions:
...
  models.BottleResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      open_date:
        type: string
      opened:
        type: boolean
      purchase_date:
        type: string
    type: object
  models.CreateBottleRequest:
    properties:
      name:
        type: string
      open_date:
        type: string
      opened:
        type: boolean
      purchase_date:
        type: string
    type: object
...
  models.UpdateBottleRequest:
    properties:
      name:
        type: string
      open_date:
        type: string
      opened:
        type: boolean
      purchase_date:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/bottles:
    get:
      description: Returns a list of all bottles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BottleResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all bottles
      tags:
      - bottles
    post:
      consumes:
      - application/json
      description: Adds a new bottle to the collection
      parameters:
      - description: Bottle to add
        in: body
        name: bottle
        required: true
        schema:
          $ref: '#/definitions/models.CreateBottleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BottleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new bottle
      tags:
      - bottles
  /api/bottles/{id}:
    delete:
      description: Deletes a bottle from the collection by its ID
      parameters:
      - description: Bottle ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a bottle by ID
      tags:
      - bottles
    get:
      description: Returns a single bottle by its ID
      parameters:
      - description: Bottle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BottleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a bottle by ID
      tags:
      - bottles
    put:
      consumes:
      - application/json
      description: Updates a bottle's information by its ID
      parameters:
      - description: Bottle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bottle update info
        in: body
        name: bottle
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBottleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BottleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a bottle by ID
      tags:
      - bottles
...
swagger: "2.0"